Add files via upload
Okay, so a few changes here.
DominoPiece: made the class inherit Dominoes to make use of methods smoother in CPlayer, as well as made left and right values protected instead of private to allow easier access for the CPlayer class to compare left and right domino sides. Also changed the "flip" method from a void method with no arguments to a method that takes and returns a DominoPiece object with a flipped left and right side.
CPlayer: finished the selectDomino method(now combination of earlier selectDomino and setDomino) which takes the int value input into the console and either sets the given domino if it has a number that correlates with the last set domino, or it does not. Changed the turn order randomization values from 0/1 to 1/-1, so that the values can be flipped with a simple *-1 operation.
Next up:  implementing an automatic check function that checks all left and right values of a person's hand to see if they can play a domino or if they need to take one from the pile.