Add files via upload
changed player turn values to boolean values and randomized those at the beginning. Added a printed layout of the currently played dominoes. SelectDomino method now adds the selected domino to the list of played dominoes. Next up: Adding a win condition, add a "check statement" for the player to check their hand, the stock, or the last played piece(might want to implement this in CTable instead given that class's nature)